+++ lib/rdf/triple.ex
@@ -87,14 +87,10 @@ defmodule RDF.Triple do
   @doc """
   Returns a tuple of native Elixir values from a `RDF.Triple` of RDF terms.
 
-  Returns `nil` if one of the components of the given tuple is not convertible via `RDF.Term.value/1`.
+  When the optional `property_map` argument is given, predicates will be mapped
+  to the terms defined in the `RDF.PropertyMap` if present.
 
-  The optional second argument allows to specify a custom mapping with a function
-  which will receive a tuple `{statement_position, rdf_term}` where
-  `statement_position` is one of the atoms `:subject`, `:predicate` or `:object`,
-  while `rdf_term` is the RDF term to be mapped. When the given function returns
-  `nil` this will be interpreted as an error and will become the overhaul result
-  of the `values/2` call.
+  Returns `nil` if one of the components of the given tuple is not convertible via `RDF.Term.value/1`.
 
   ## Examples
 
@@ -102,27 +98,53 @@ defmodule RDF.Triple do
       {"http://example.com/S", "http://example.com/p", 42}
 
       iex> {~I<http://example.com/S>, ~I<http://example.com/p>, RDF.literal(42)}
-      ...> |> RDF.Triple.values(fn
+      ...> |> RDF.Triple.values(PropertyMap.new(p: ~I<http://example.com/p>))
+      {"http://example.com/S", :p, 42}
+
+
+  """
+  @spec values(t, PropertyMap.t() | nil) :: t_values | nil
+  def values(triple, property_map \\ nil)
+
+  def values(triple, nil) do
+    map(triple, &Statement.default_term_mapping/1)
+  end
+
+  def values(triple, %PropertyMap{} = property_map) do
+    map(triple, Statement.default_property_mapping(property_map))
+  end
+
+  @doc """
+  Returns a triple where each element from a `RDF.Triple` is mapped with the given function.
+
+  Returns `nil` if one of the components of the given tuple is not convertible via `RDF.Term.value/1`.
+
+  The function `fun` will receive a tuple `{statement_position, rdf_term}` where
+  `statement_position` is one of the atoms `:subject`, `:predicate` or `:object`,
+  while `rdf_term` is the RDF term to be mapped. When the given function returns
+  `nil` this will be interpreted as an error and will become the overhaul result
+  of the `map/2` call.
+
+  ## Examples
+
+      iex> {~I<http://example.com/S>, ~I<http://example.com/p>, RDF.literal(42)}
+      ...> |> RDF.Triple.map(fn
       ...>      {:object, object} -> RDF.Term.value(object)
       ...>      {_, term}         -> term |> to_string() |> String.last()
       ...>    end)
       {"S", "p", 42}
 
   """
-  @spec values(t | any, Statement.term_mapping()) :: t_values | nil
-  def values(triple, mapping \\ &Statement.default_term_mapping/1)
-
-  def values({subject, predicate, object}, mapping) do
-    with subject_value when not is_nil(subject_value) <- mapping.({:subject, subject}),
-         predicate_value when not is_nil(predicate_value) <- mapping.({:predicate, predicate}),
-         object_value when not is_nil(object_value) <- mapping.({:object, object}) do
+  @spec map(t, Statement.term_mapping()) :: t_values | nil
+  def map({subject, predicate, object}, fun) do
+    with subject_value when not is_nil(subject_value) <- fun.({:subject, subject}),
+         predicate_value when not is_nil(predicate_value) <- fun.({:predicate, predicate}),
+         object_value when not is_nil(object_value) <- fun.({:object, object}) do
       {subject_value, predicate_value, object_value}
     else
       _ -> nil
     end
   end
-
-  def values(_, _), do: nil
 
   @doc """
   Checks if the given tuple is a valid RDF triple.