+++ test/unit/triple_test.exs
@@ -12,14 +12,33 @@ defmodule RDF.TripleTest do
     end
 
     test "with an invalid RDF.Triple" do
-      refute Triple.values({~I<http://example.com/S>, ~I<http://example.com/p>})
       refute Triple.values({self(), self(), self()})
     end
   end
 
-  test "values/2" do
+  describe "values/2" do
+    test "with a valid RDF.Triple and RDF.PropertyMap" do
+      assert Triple.values(
+               {~I<http://example.com/S>, ~I<http://example.com/p>, XSD.integer(42)},
+               PropertyMap.new(p: ~I<http://example.com/p>)
+             ) ==
+               {"http://example.com/S", :p, 42}
+
+      assert Triple.values(
+               {~I<http://example.com/S>, ~I<http://example.com/p>, XSD.integer(42)},
+               PropertyMap.new()
+             ) ==
+               {"http://example.com/S", "http://example.com/p", 42}
+    end
+
+    test "with an invalid RDF.Triple" do
+      refute Triple.values({self(), self(), self()}, PropertyMap.new())
+    end
+  end
+
+  test "map/2" do
     assert {~I<http://example.com/S>, ~I<http://example.com/p>, XSD.integer(42)}
-           |> Triple.values(fn
+           |> Triple.map(fn
              {:object, object} -> object |> RDF.Term.value() |> Kernel.+(1)
              {_, term} -> term |> to_string() |> String.last()
            end) ==